feat(SideBar): Add selected count and clear button
